Break out command popper from message component
